requirements.txt
streamlit
openai
python-pptx
pillow
requests
